reverted back to old code
interestingly a switch with a case for each possible scenario was worse than the teired approach initially used. This is because when it needs to pull a case from the bottom of the script it tries every single case before the one that works. What this amounted to was a performance decrease of roughly 50% across the board